+++ apps/tenant-selector/src/sidecar.ts
@@ -1,4 +1,4 @@
-import { SidecarConfig } from '@cognite/react-tenant-selector';
+import { SidecarConfig } from '@cognite/sidecar';
 
 export const SIDECAR: SidecarConfig = {
   // eslint-disable-next-line no-underscore-dangle