+++ packages/react-container/src/components/__tests__/Intercom.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import * as intercomPackage from '@cognite/intercom-helper';
 import { act } from 'react-dom/test-utils';
-import { generateSidecar } from '__mocks/sidecar';
+import { mockSidecar } from '@cognite/sidecar';
 
 import { IntercomContainer } from '../Intercom';
 
@@ -23,7 +23,7 @@ describe('Intercom', () => {
   });
 
   it('should render', async () => {
-    const sidecar = generateSidecar();
+    const sidecar = mockSidecar();
     const intercomSettings = {
       app_id: 'myId',
       hide_default_launcher: true,
@@ -63,7 +63,7 @@ describe('Intercom', () => {
   });
 
   it('should not render', () => {
-    const sidecar = generateSidecar();
+    const sidecar = mockSidecar();
 
     const Test = () => (
       <IntercomContainer sidecar={sidecar} project="">