+++ packages/react-tenant-selector/src/index.ts
@@ -1,2 +1,1 @@
 export * from './TenantSelector';
-export type { SidecarConfig } from './utils';