+++ packages/react-tenant-selector/src/utils/index.ts
@@ -1,3 +1,2 @@
 export { default as config } from './config';
-export type { SidecarConfig, FakeIdp } from './types';
 export * from './tenancy';