+++ packages/sidecar/src/index.ts
@@ -1,2 +1,3 @@
 export * from './getDefaultSidecar';
 export type { CDFCluster, FakeIdp, SidecarConfig } from './types';
+export * from './mocks';